+++ packages/plugins/other/visitor-plugin-common/src/types.ts
@@ -1,4 +1,4 @@
-import { ASTNode, FragmentDefinitionNode } from 'graphql';
+import { ASTNode, FragmentDefinitionNode, DirectiveNode } from 'graphql';
 import { ParsedMapper } from './mappers.js';
 
 /**
@@ -103,6 +103,12 @@ export interface ParsedImport {
   propName: string;
 }
 
+export type FragmentDirectives = {
+  fragmentDirectives?: Array<DirectiveNode>;
+};
+
 export interface ResolversNonOptionalTypenameConfig {
   unionMember?: boolean;
+  interfaceImplementingType?: boolean;
+  excludeTypes?: string[];
 }