+++ packages/presets/client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @graphql-codegen/client-preset
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`386cf9044`](https://github.com/dotansimha/graphql-code-generator/commit/386cf9044a41d87ed45069b22d26b30f4b262a85), [`402cb8ac0`](https://github.com/dotansimha/graphql-code-generator/commit/402cb8ac0f0c347b186d295c4b69c19e25a65d00)]:
+  - @graphql-codegen/visitor-plugin-common@3.1.1
+  - @graphql-codegen/gql-tag-operations@3.0.1
+  - @graphql-codegen/typescript-operations@3.0.4
+  - @graphql-codegen/typed-document-node@4.0.1
+  - @graphql-codegen/typescript@3.0.4
+
 ## 3.0.0
 
 ### Major Changes